+++ R/metrics_crowns.R
@@ -38,7 +38,23 @@ crown_metrics.LAS = function(las, func, geom = "point", concaveman = c(3, 0), at
 
   M1 <- las@data[, if (!anyNA(.BY)) fgeom(X,Y,Z, concavity, length_threshold), by = .(GRPID = template)]
   data.table::setorder(M1, GRPID)
-  geom <- sf::st_as_sfc(M1[["geom"]])
+  sfgeom <- sf::st_as_sfc(M1[["geom"]])
+
+
+  if (is(sfgeom, "sfc_POLYGON"))
+  {
+    invalid <- !sf::st_is_valid(sfgeom)
+    ninvalid <- sum(invalid)
+
+    if (ninvalid > 0)
+    {
+      if (geom == "concave")
+        warning(glue::glue("{ninvalid} invalid polygons created. They likely correspond either to trees with aligned points or to edge cases where the convex hull converged to polygons that are not valid."), call. = FALSE)
+
+      if (geom == "convex")
+        warning(glue::glue("{ninvalid} invalid polygons created. They likely correspond to trees with aligned points."), call. = FALSE)
+    }
+  }
 
   if (!is.null(func))
   {
@@ -69,7 +85,7 @@ crown_metrics.LAS = function(las, func, geom = "point", concaveman = c(3, 0), at
     M2 <- cbind(M2, M3)
   }
 
-  output <- sf::st_set_geometry(M2, geom)
+  output <- sf::st_set_geometry(M2, sfgeom)
   sf::st_crs(output) <- st_crs(las)
   return(output)
 }